Deal with note de-bouncing
Unfortunately, it doesn't look like tk provides any information on whether a
keypress event is the result of a physical keypress or an auto-repeat, so we
need to manage this in the code.

Deal with it by keeping track of the last release time and deferring
processing until 100th of a second after that release (assuming no subsequent
keypress at the same time has occurred).